# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zeitgeist",
    "version": "0.1.0"
  },
  "paths": {
    "/api/runs": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Start Run",
        "description": "202 rather than 201: the run is queued, and for a queued one there is\nnothing at its URL yet.",
        "operationId": "start_run_api_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/events": {
      "get": {
        "tags": [
          "control"
        ],
        "summary": "Stream Events",
        "description": "Log lines and progress ticks, until the run ends.\n\nPolls the run's buffer rather than being pushed to. A quarter-second of\nlatency on a log line is invisible, and polling keeps the logging handler\nfree of any reference to the event loop \u2014 which is what lets the same\nhandler work under `TestClient` and under a real server alike.\n\nA tick carries no payload. The observer has already written\n`run_records`, `run_stages` and `run_topics` from the worker thread, so\nthe client refetches rather than being handed the same data twice.",
        "operationId": "stream_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Run",
        "operationId": "read_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/topics": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Ranking",
        "operationId": "read_ranking_api_runs__run_id__topics_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RankedTopic"
                  },
                  "title": "Response Read Ranking Api Runs  Run Id  Topics Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/topics/{topic_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Topic",
        "operationId": "read_topic_api_runs__run_id__topics__topic_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/log": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Log",
        "description": "One run's whole log, every time.\n\nNo `since_seq`. Phase 2 left the question open because incremental\npolling was the only way a live log could have worked; phase 3's\n`GET /api/runs/{id}/events` settled it by making this endpoint the\n*other* half of a pair. The stream carries a live run's lines, keyed on\nexactly the `seq` a `since_seq` here would have taken, and the client\nenables the two exclusively \u2014 `RunDetailPage` opens the stream while\nthe run's status is `running` and only calls this once it is not.\n\nSo the one caller asks a finished run for its whole log, once, and a\n`since_seq` would be a parameter nothing passes: a second cursor\nprotocol, a second set of tests, and a second way for the log to arrive\nwith a hole in it. A finished run's log is bounded by the run, and the\nverbose toggle filters what is *returned* rather than what was\ncaptured, which is what lets flipping it work retroactively \u2014 that only\nholds while the endpoint serves the whole thing.",
        "operationId": "read_log_api_runs__run_id__log_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "verbose",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Verbose"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogLine"
                  },
                  "title": "Response Read Log Api Runs  Run Id  Log Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogLine": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "logged_at": {
            "type": "string",
            "format": "date-time",
            "title": "Logged At"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "logger": {
            "type": "string",
            "title": "Logger"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "seq",
          "logged_at",
          "level",
          "logger",
          "message"
        ],
        "title": "LogLine",
        "description": "One line of a run's log.\n\n`seq` orders lines within a run: two lines can share a timestamp at the\nresolution the handler records, and the live log's ordering has to be\ntotal."
      },
...
      "QueuedRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "position"
        ],
        "title": "QueuedRun",
        "description": "`position` 0 means executing now; higher means waiting behind that\nmany runs, which is what the \"queues it behind\" notice reads."
      },
      "RankedTopic": {
        "properties": {
          "topic": {
            "$ref": "#/components/schemas/TopicRow"
          },
          "render_count": {
            "type": "integer",
            "title": "Render Count"
          },
          "above_cut": {
            "type": "boolean",
            "title": "Above Cut"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "topic",
          "render_count",
          "above_cut"
        ],
        "title": "RankedTopic",
        "description": "One row of the ranking list.\n\n`above_cut` is computed from the run's frozen `top_count` rather than\nleft to the client: the screen draws rows below the cut dashed and\ndimmed with a generate link, so the rule is load-bearing."
      },
...
      "ReplyOut": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "like_count",
          "created_at"
        ],
        "title": "ReplyOut",
        "description": "One reply, as the dossier shows it.\n\nNo author, handle or avatar. `Reply.author_key` exists only to count the\ndistinct accounts behind a repeated phrase; the project stores no\npersonal data and an API that returned it would undo that."
      },
...
      "RunDetail": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunRecordRow"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageRecord"
            },
            "type": "array",
            "title": "Stages"
          },
          "resume_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Stage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run",
          "stages",
          "resume_stage"
        ],
        "title": "RunDetail",
        "description": "One run's header: what it was configured with, how its stages went,\nand where a resume would start."
      },
...
      "RunPage": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runs",
          "next_cursor"
        ],
        "title": "RunPage",
        "description": "One page of runs. `next_cursor` is None on the last page."
      },
      "RunRecordRow": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "ok",
              "failed",
              "aborted",
              "interrupted"
            ],
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "attempt_started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Attempt Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "config": {
            "$ref": "#/components/schemas/RunConfig"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunError"
              },
              {
                "type": "null"
              }
            ]
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "trends_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trends Found"
          },
          "topics_kept": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topics Kept"
          },
          "phrases_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phrases Found"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "started_at",
          "attempt_started_at",
          "finished_at",
          "config",
          "error",
          "item_count",
          "trends_found",
          "topics_kept",
          "phrases_found"
        ],
        "title": "RunRecordRow",
        "description": "A row of `run_records`, read back.\n\nThe counts are `| None` because they are written when the run finishes;\na run still in flight has not counted anything yet."
      },
      "RunSummary": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunRecordRow"
          },
          "topic_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topic Labels"
          },
          "render_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Render Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run",
          "topic_labels",
          "render_ids"
        ],
        "title": "RunSummary",
        "description": "One row of the Runs list.\n\n`topic_labels` and `render_ids` are joins the row needs and the run\nrecord does not carry: column two is the titles, column four the\nthumbnails."
      },
...
      "Stage": {
        "type": "string",
        "enum": [
          "ingest",
          "analyse",
          "evaluate",
          "generate"
        ],
        "title": "Stage"
      },
      "StageRecord": {
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/Stage"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "ok",
              "failed",
              "skipped"
            ],
            "title": "Status"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "payload_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Bytes"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "done": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Done"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "stage",
          "status",
          "started_at",
          "finished_at",
          "payload_bytes",
          "summary"
        ],
        "title": "StageRecord",
        "description": "One stage of one run.\n\nEvery `| None` here is a real state: a queued stage has not started, a\nrunning one has not finished, and a failed or skipped one wrote no\ncheckpoint.\n\n`done` and `total` are the stage-relative counters the in-flight card\ndraws as `17 / 25`, and the partial fill on its top bar. They default\nto None because two of the four stages have nothing to count \u2014 ingest\nis one opaque fetch and evaluate one ranking pass \u2014 so requiring every\nconstruction site to spell out `done=None, total=None` would state\nnothing four times over. A running stage that reports no counters\nrenders an indeterminate bar rather than a zero-length one."
      },
      "StartRunBody": {
        "properties": {
          "template_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Ids"
          },
          "overrides": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Overrides",
            "default": {}
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "StartRunBody",
        "description": "What `POST /api/runs` accepts.\n\n`overrides` is `str`-valued because the values come from a form and are\nvalidated by `Settings` when the worker freezes the run's config \u2014 the\none place that knows each field's real type."
      },
...
      "TopicDetail": {
        "properties": {
          "topic": {
            "$ref": "#/components/schemas/TopicRow"
          },
          "dossier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Dossier"
              },
              {
                "type": "null"
              }
            ]
          },
          "score_components": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Score Components"
          },
          "replies": {
            "items": {
              "$ref": "#/components/schemas/ReplyOut"
            },
            "type": "array",
            "title": "Replies"
          },
          "renders": {
            "items": {
              "$ref": "#/components/schemas/RenderRecord"
            },
            "type": "array",
            "title": "Renders"
          },
          "recurrence": {
            "$ref": "#/components/schemas/TopicRecurrence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "topic",
          "dossier",
          "score_components",
          "replies",
          "renders",
          "recurrence"
        ],
        "title": "TopicDetail",
        "description": "One topic's dossier: the row, the prose, what people said, and what\nwas rendered from it."
      },
...
      "TopicRow": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "topic_id": {
            "type": "string",
            "title": "Topic Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "label_slug": {
            "type": "string",
            "title": "Label Slug"
          },
          "trend_status": {
            "type": "string",
            "enum": [
              "trending",
              "saturating",
              "cooling",
              "stale"
            ],
            "title": "Trend Status"
          },
          "event_sentiment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Sentiment"
          },
          "conversation_register": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Register"
          },
          "meme_potential": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meme Potential"
          },
          "trend_score": {
            "type": "number",
            "title": "Trend Score"
          },
          "final_score": {
            "type": "number",
            "title": "Final Score"
          },
          "final_rank": {
            "type": "integer",
            "title": "Final Rank"
          },
          "post_count": {
            "type": "integer",
            "title": "Post Count"
          },
          "top_phrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Phrase"
          },
          "top_phrase_authors": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Phrase Authors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "topic_id",
          "label",
          "label_slug",
          "trend_status",
          "event_sentiment",
          "conversation_register",
          "meme_potential",
          "trend_score",
          "final_score",
          "final_rank",
          "post_count",
          "top_phrase",
          "top_phrase_authors"
        ],
        "title": "TopicRow",
        "description": "One row of `run_topics`."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}